Moved task and job management to separate files
