+++ support.js
@@ -2,6 +2,7 @@ import { querySudo as query, updateSudo as update } from '@lblod/mu-auth-sudo';
 import { uuid, sparqlEscapeString, sparqlEscapeDateTime, sparqlEscapeUri } from 'mu';
 import { Writer } from 'n3';
 import * as env from './env.js';
+import * as jobsAndTasks from './jobAndTaskManagement.js';
 
 async function isSubmitted(resource) {
   const result = await query(`
@@ -55,276 +56,10 @@ async function triplesToTurtle(triples) {
   return promise;
 }
 
-async function startJob(submissionGraph, meldingUri) {
-  const jobUuid = uuid();
-  const nowSparql = sparqlEscapeDateTime((new Date()).toISOString());
-  // Make a cogs:Job for the whole process
-  // The prov:generated is strictly not necessary for the model, maybe nice to have
-  const jobQuery = `
-    ${PREFIXES}
-    INSERT DATA {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        asj:${jobUuid}
-          a cogs:Job ;
-          mu:uuid ${sparqlEscapeString(jobUuid)} ;
-          dct:creator services:automatic-submission-service ;
-          adms:status js:busy ;
-          dct:created ${nowSparql} ;
-          dct:modified ${nowSparql} ;
-          task:operation cogs:TransformationProcess ;
-          prov:generated ${sparqlEscapeUri(meldingUri)} .
-      }
-    }
-  `;
-  await update(jobQuery);
-
-  // Create a task for the automatic submission as the first step in the flow
-  const submissionTaskUuid = uuid();
-  const submissionTaskQuery = `
-    ${PREFIXES}
-    INSERT DATA {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        asj:${submissionTaskUuid}
-          a task:Task ;
-          mu:uuid ${sparqlEscapeString(submissionTaskUuid)} ;
-          adms:status js:busy ;
-          dct:created ${nowSparql} ;
-          dct:modified ${nowSparql} ;
-          task:operation cogs:TransformationProcess ;
-          dct:creator services:automatic-submission-service ;
-          task:index "0" ;
-          dct:isPartOf asj:${jobUuid} .
-      }
-    }
-  `;
-  await update(submissionTaskQuery);
-
-  const jobUri = `http://data.lblod.info/id/automatic-submission-job/${jobUuid}`
-  const automaticSubmissionTaskUri = `http://data.lblod.info/id/automatic-submission-job/${submissionTaskUuid}`;
-  return { jobUri, automaticSubmissionTaskUri, };
-}
-
-async function automaticSubmissionTaskSuccess(submissionGraph, automaticSubmissionTaskUri, jobUri) {
-  const nowSparql = sparqlEscapeDateTime((new Date()).toISOString());
-  const automaticSubmissionTaskUriSparql = sparqlEscapeUri(automaticSubmissionTaskUri);
-  const resultContainerUuid = uuid();
-  const assTaskQuery = `
-    ${PREFIXES}
-    DELETE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${automaticSubmissionTaskUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-    INSERT {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${automaticSubmissionTaskUriSparql}
-          adms:status js:success ;
-          dct:modified ${nowSparql} ;
-          task:resultsContainer asj:${resultContainerUuid} .
-
-        asj:${resultContainerUuid}
-          a nfo:DataContainer ;
-          mu:uuid ${sparqlEscapeString(resultContainerUuid)} .
-      }
-    }
-    WHERE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${automaticSubmissionTaskUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-  `;
-  await update(assTaskQuery);
-
-  const downloadTaskUuid = uuid();
-  const inputContainerUuid = uuid();
-  const downloadTaskQuery = `
-    ${PREFIXES}
-    INSERT DATA {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        asj:${downloadTaskUuid}
-          a task:Task ;
-          mu:uuid ${sparqlEscapeString(downloadTaskUuid)} ;
-          adms:status js:busy ;
-          dct:created ${nowSparql} ;
-          dct:modified ${nowSparql} ;
-          task:operation cogs:WebServiceLookup ;
-          dct:creator services:automatic-submission-service ;
-          task:index "1" ;
-          dct:isPartOf ${sparqlEscapeUri(jobUri)} ;
-          task:inputContainer asj:${inputContainerUuid} .
-
-        asj:j${inputContainerUuid}
-          a nfo:DataContainer ;
-          mu:uuid ${sparqlEscapeString(inputContainerUuid)} .
-      }
-    }
-  `;
-  await update(downloadTaskQuery);
-  
-  const downloadTaskUri = `http://data.lblod.info/id/automatic-submission-job/${downloadTaskUuid}`;
-  return downloadTaskUri;
-}
-
-async function automaticSubmissionTaskFail(submissionGraph, automaticSubmissionTaskUri, jobUri) {
-  const nowSparql = sparqlEscapeDateTime((new Date()).toISOString());
-  const automaticSubmissionTaskUriSparql = sparqlEscapeUri(automaticSubmissionTaskUri);
-  const assTaskQuery = `
-    ${PREFIXES}
-    DELETE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${automaticSubmissionTaskUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-    INSERT {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${automaticSubmissionTaskUriSparql}
-          adms:status js:failed ;
-          dct:modified ${nowSparql} .
-      }
-    }
-    WHERE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${automaticSubmissionTaskUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-  `;
-  await update(assTaskQuery);
-
-  //Also set the job to failure
-  const jobUriSparql = sparqlEscapeUri(jobUri);
-  const assJobQuery = `
-    ${PREFIXES}
-    DELETE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${jobUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-    INSERT {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${jobUriSparql}
-          adms:status js:failed ;
-          dct:modified ${nowSparql} .
-      }
-    }
-    WHERE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${jobUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-  `;
-  await update(assJobQuery);
-}
-
-async function downloadSuccess(submissionGraph, downloadTaskUri) {
-  const nowSparql = sparqlEscapeDateTime((new Date()).toISOString());
-  const resultContainerUuid = uuid();
-  const downloadTaskUriSparql = sparqlEscapeUri(downloadTaskUri);
-  const downloadTaskQuery = `
-    ${PREFIXES}
-    DELETE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${downloadTaskUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-    INSERT {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${downloadTaskUriSparql}
-          adms:status js:success ;
-          dct:modified ${nowSparql} ;
-          task:resultsContainer asj:${resultContainerUuid} .
-
-        asj:${resultContainerUuid}
-          a nfo:DataContainer ;
-          mu:uuid ${sparqlEscapeString(resultContainerUuid)} .
-      }
-    }
-    WHERE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${downloadTaskUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-  `;
-  await update(downloadTaskQuery);
-}
-
-async function downloadFail(submissionGraph, downloadTaskUri, jobUri) {
-  const nowSparql = sparqlEscapeDateTime((new Date()).toISOString());
-  const downloadTaskUriSparql = sparqlEscapeUri(downloadTaskUri);
-  const downloadTaskQuery = `
-    ${PREFIXES}
-    DELETE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${downloadTaskUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-    INSERT {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${downloadTaskUriSparql}
-          adms:status js:failed ;
-          dct:modified ${nowSparql} .
-      }
-    }
-    WHERE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${downloadTaskUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-  `;
-  await update(downloadTaskQuery);
-
-  //Also set the job to failure
-  const jobUriSparql = sparqlEscapeUri(jobUri);
-  const assJobQuery = `
-    ${PREFIXES}
-    DELETE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${jobUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-    INSERT {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${jobUriSparql}
-          adms:status js:failed ;
-          dct:modified ${nowSparql} .
-      }
-    }
-    WHERE {
-      GRAPH ${sparqlEscapeUri(submissionGraph)} {
-        ${jobUriSparql}
-          adms:status ?oldStatus ;
-          dct:modified ?oldModified .
-      }
-    }
-  `;
-  await update(assJobQuery);
-}
-
 async function storeSubmission(triples, submissionGraph, fileGraph, authenticationConfiguration) {
   let newAuthConf = {};
   const meldingUri = extractMeldingUri(triples);
-  const { jobUri, automaticSubmissionTaskUri, } = await startJob(submissionGraph, meldingUri);
+  const { jobUri, automaticSubmissionTaskUri, } = await jobsAndTasks.startJob(submissionGraph, meldingUri);
   try {
     const submittedResource = findSubmittedResource(triples);
     const turtle = await triplesToTurtle(triples);
@@ -400,14 +135,14 @@ async function storeSubmission(triples, submissionGraph, fileGraph, authenticati
       }
     `);
 
-    await automaticSubmissionTaskSuccess(submissionGraph, automaticSubmissionTaskUri, jobUri);
+    await jobsAndTasks.automaticSubmissionTaskSuccess(submissionGraph, automaticSubmissionTaskUri, jobUri);
 
     return jobUri;
   } catch (e) {
     console.error('Something went wrong during the storage of submission');
     console.error(e);
     console.info('Cleaning credentials');
-    await automaticSubmissionTaskFail(submissionGraph, automaticSubmissionTaskUri, jobUri);
+    await jobsAndTasks.automaticSubmissionTaskFail(submissionGraph, automaticSubmissionTaskUri, jobUri);
     await cleanCredentials(authenticationConfiguration);
     if (newAuthConf.newAuthConf) {
       await cleanCredentials(newAuthConf.newAuthConf);
